requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
ca-certificates=2024.3.11=h06a4308_0
contourpy=1.1.1=pypi_0
cycler=0.12.1=pypi_0
filelock=3.13.4=pypi_0
fonttools=4.51.0=pypi_0
fsspec=2024.3.1=pypi_0
importlib-resources=6.4.0=pypi_0
jinja2=3.1.3=pypi_0
kiwisolver=1.4.5=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
markupsafe=2.1.5=pypi_0
matplotlib=3.7.5=pypi_0
mpmath=1.3.0=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.1=pypi_0
numpy=1.24.4=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.19.3=pypi_0
nvidia-nvjitlink-cu12=12.4.127=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
opencv-python=4.9.0.80=pypi_0
openssl=3.0.13=h7f8727e_0
packaging=24.0=pypi_0
pillow=10.3.0=pypi_0
pip=23.3.1=py38h06a4308_0
pyparsing=3.1.2=pypi_0
python=3.8.19=h955ad1f_0
python-dateutil=2.9.0.post0=pypi_0
readline=8.2=h5eee18b_0
setuptools=68.2.2=py38h06a4308_0
six=1.16.0=pypi_0
sqlite=3.41.2=h5eee18b_0
sympy=1.12=pypi_0
tk=8.6.12=h1ccaba5_0
torch=2.2.2=pypi_0
torchvision=0.17.2=pypi_0
tqdm=4.66.2=pypi_0
triton=2.2.0=pypi_0
typing-extensions=4.11.0=pypi_0
wheel=0.41.2=py38h06a4308_0
xz=5.4.6=h5eee18b_0
zipp=3.18.1=pypi_0
zlib=1.2.13=h5eee18b_0
